tinfernew/requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot
httpx
